Restrict certain processes to limited CPUs when non-interactive
By using cgroup/cpuset, certain processes are now restricted
to running on a limited number of CPU cores when userspace deems
device to be non-interactive. This helps with power consumption
during screen off, as the other CPU cores can be kept at
low power state.

A helper program is introduced to look for PIDs for interesting
processes. The help sets a property with each PID and let
init.rc to take care of writing to the cgroup sysfs nodes.
This is because power HAL is running under user system, and
kernel does not allow changing scheduling parameters for
processed running under different user. Since init is running
under root, the kernel does not block that operation.

Initially, this is used to control mediaserver, for better
power consumption on screen off audio playback.

Change-Id: I3f10f9821c57204bb72c734f3ef8adb3bafb3a80
Signed-off-by: Daniel Leung <daniel.leung@intel.com>